This what we saw on 19 july class, we talked about navigation and using fetch and axios for getting api information.
We installed react native navigation and its set up for android and ios use. Also is added in the package.json so just run yarn install and it will be there.

Sorry for those bunch of files that we added on this commit, but like this was a new project it would create all from zero >.< !